+++ MoebiusEngine/src/main/java/civitas/celestis/math/matrix/IntMatrix.java
@@ -491,7 +491,7 @@ public class IntMatrix implements Matrix<Integer, IntMatrix> {
 
         for (int r = 0; r < rows; r++) {
             for (int c = 0; c < columns; c++) {
-                result.values[r][c] = (int) f.apply(values[r][c]);
+                result.values[r][c] = f.apply(values[r][c]);
             }
         }
 
@@ -676,7 +676,7 @@ public class IntMatrix implements Matrix<Integer, IntMatrix> {
      *
      * @param s String to parse
      * @return Matrix parsed from given string
-     * @throws NumberFormatException    When the string is not parsable to a matrix
+     * @throws NumberFormatException When the string is not parsable to a matrix
      */
     @Nonnull
     public static IntMatrix parseMatrix(@Nonnull String s) throws NumberFormatException {
